Prevent duplicate user registration by blocking signups with existing email addresses across all providers.
If the user tries to create a credentialed account after loging in with a provider with the same email address, such as google, don't allow it.

However, if the user creates a credentialed account first, then logs in with a provider with the same email address, allow it.

This essentially extends a credentialed user so that with the same credentials, the application can access specific APIs like Google calendar, for instance.